+++ lib/Entry.php
