+++ lib/config/FrontVariables.php
