+++ lib/modules/Widget.php
