+++ lib/modules/clean.php
