+++ lib/plugins/Template/Kiku-base.php
