+++ lib/plugins/Template/Kiku-loader.php
