+++ lib/plugins/mokuji/Mokuji.php
